+++ ostap/fitting/variables.py
@@ -24,9 +24,11 @@ __all__     = (
     'total_ratio'     , ## ``converter'': A,B ->  (T,R) == ( A+B , A/B     )
     'total_ratio'     , ## ``converter'': A,B ->  (T,F) == ( A+B , A/(A+B) ) 
     'two_yields'      , ## ``converter'': T,F ->  (A,B) == ( R*F , T*(1-F) )
+    'depends_on'      , ## Is this "RooFit" function depends on the variable?
+    'binning'         , ## create RooBinning object 
     ) 
 # =============================================================================
-import ROOT, random
+import ROOT, random, array 
 from   ostap.core.core  import VE, hID 
 from   ostap.core.ostap_types import ( num_types     , list_types   ,
                                        integer_types , string_types )   
@@ -1046,7 +1048,66 @@ ROOT.RooUniformBinning.__str__  = _rub_str_
 ROOT.RooUniformBinning.__repr__ = _rub_str_
 
 
+# =============================================================================
+## Does  this variable depends on another one?
+#  @code
+#  fun = ...
+#  var = ...
+#  if fun.depends_on ( var ) :
+#     ...
+#  @endcode
+def depends_on ( fun , var ) :
+    """Does  this variable depends on another one?
+    
+    >>> fun = ...
+    >>> var = ...
+    >>> if fun.depends_on ( var ) :
+    ...
+    
+    """
+    if isinstance ( var , ROOT.RooAbsCollection ) :
+        for v in var :
+            if depends_on ( fun , v ) : return True
+        return False
 
+    fpars = fun.getParameters ( 0 )
+    
+    ## direct dependency?
+    if var  in fpars : return True
+        
+    ## check indirect dependency
+    vvars = var.getParameters ( 0 )
+    for v in vvars :
+        if v in fpars : return True
+
+    ##
+    return False 
+
+ROOT.RooAbsReal.depends_on  = depends_on
+
+# =============================================================================
+## Create <code>RooBinnig</code> object
+#  @param edges vector of bin edges 
+#  @param nbins number of bins
+#  @param name  binning name
+#  @see RooBinning
+def binning ( edges , nbins = 0 , name = '' ) :
+    """Create `RooBinnig` object
+    - see ROOT.RooBinning
+    """
+    assert isinstance ( nbins , integer_types ) and 0 <= nbins, \
+           "Invalid ``nbins'' parameter %s/%s" % ( nbins , type ( nbins ) )
+
+    nb = len ( edges ) 
+    assert 2 <= nb , "Invalid length of ``edges'' array!"
+
+    if 2 == nb :
+        return ROOT.RooBinning ( max ( 1 , nbins ) , edges[0] , edges[1] , name ) 
+
+    buffer = array.array ( 'd', edges )
+    return ROOT.RooBinning ( nb - 1 , bufffer , name ) 
+    
+    
 # =============================================================================
 _decorated_classes_ = (
     ROOT.RooRealVar        ,